+++ CHANGELOG.md
@@ -7,6 +7,10 @@ is a shorter change log with only the SAU language changes.
 Pre-release
 -----------
 
+Language changes:
+ * Line types. More exp/log types with steepness not 6.
+    - Add `exp11`, `log11`, `xpe11`, `lge11`. Steepness 11.
+
 v0.5.1 (2025-02-16)
 -------------------